Move location conversions to mapper
